Buttons for needed steps set to 20 or 10 atm and when you have enough steps it lets you go to the next chapter. Also made so that if you have gone through the chapter once you don't need to play it again. Might have some bugs still.

+++ core/src/fi/tamk/tiko/ChapterAssets.java
@@ -43,7 +43,7 @@ public class ChapterAssets extends Actor {
     float screenHeight;
 
     float textboxHeight;
-    float steps;
+    int steps;
     String currentStepsText;
 
     public ChapterAssets(MainClass MainClass2, Texture texture, boolean choice, int affect_id, int affect_steps, String choiceBox1, String choiceBox2, int stepCount, String story) {
